Provide code coverage with unit tests & refactoring of the executor service

+++ problem-service/src/main/java/com/danyazero/problemservice/service/ProblemService.java
@@ -1,52 +1,60 @@
 package com.danyazero.problemservice.service;
 
 import com.danyazero.problemservice.entity.Problem;
+import com.danyazero.problemservice.exception.IllegalRequstArgumentException;
 import com.danyazero.problemservice.exception.RequestException;
 import com.danyazero.problemservice.model.CreateProblemDto;
 import com.danyazero.problemservice.model.PageDto;
 import com.danyazero.problemservice.model.ProblemDto;
 import com.danyazero.problemservice.model.ProblemResponse;
-import com.danyazero.problemservice.model.TestcaseDto;
+import com.danyazero.problemservice.model.TestDto;
 import com.danyazero.problemservice.repository.ProblemRepository;
-import com.danyazero.problemservice.repository.TestcaseRepository;
+import com.danyazero.problemservice.repository.TestRepository;
 import com.danyazero.problemservice.utils.ProblemSearchSpecification;
 import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.data.jpa.domain.Specification;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
+@Slf4j
 @Service
 @RequiredArgsConstructor
 public class ProblemService {
 
     private final ProblemRepository problemRepository;
-    private final TestcaseRepository testcaseRepository;
+    private final TestRepository testRepository;
 
     @Transactional
     public Problem createProblem(CreateProblemDto problem) {
+        if (problem == null) {
+            log.warn("Cannot create problem, problem payload cannot be null");
+            throw new IllegalRequstArgumentException("Problem payload cannot be null");
+        }
         return problemRepository.save(problem.toEntity());
     }
 
     public ProblemResponse getProblemById(int problemId) {
         final var problem = problemRepository
             .findById(problemId)
-            .orElseThrow(() ->
-                new RequestException(
+            .orElseThrow(() -> {
+                log.warn("Problem with id {} not found.", problemId);
+                return new RequestException(
                     "Problem with id " + problemId + " not found."
-                )
-            );
+                );
+            });
 
-        final var publicTestcases = testcaseRepository
+        final var publicTests = testRepository
             .findByProblem_IdAndIsPublic(problemId, true)
             .stream()
-            .map(TestcaseDto::toDto)
+            .map(TestDto::toDto)
             .toList();
 
         return ProblemResponse.builder()
             .description(problem.getDescription())
             .difficulty(problem.getDifficulty())
-            .testcases(publicTestcases)
+            .testcases(publicTests)
             .title(problem.getTitle())
             .tags(problem.getTags())
             .id(problem.getId())
@@ -55,7 +63,11 @@ public class ProblemService {
 
     @Transactional
     public void deleteProblem(Integer problemId) {
-        testcaseRepository.deleteAllByProblem_Id(problemId);
+        if (problemId == null) {
+            log.warn("Cannot delete problem, problem ID cannot be null");
+            throw new IllegalRequstArgumentException("Problem ID cannot be null");
+        }
+        testRepository.deleteAllByProblem_Id(problemId);
         problemRepository.deleteById(problemId);
     }
 
@@ -66,6 +78,15 @@ public class ProblemService {
         int page,
         int size
     ) {
+        if (page < 0) {
+            log.warn("Cannot find problems, page index must not be less than zero");
+            throw new IllegalRequstArgumentException("Page index must not be less than zero");
+        }
+        if (size < 1) {
+            log.warn("Cannot find problems, page size must not be less than one");
+            throw new IllegalRequstArgumentException("Page size must not be less than one");
+        }
+
         Specification<Problem> spec = Specification.allOf(
             ProblemSearchSpecification.hasDifficulty(difficulty),
             ProblemSearchSpecification.hasTag(tag),
@@ -76,5 +97,53 @@ public class ProblemService {
             problemRepository.findAll(spec, PageRequest.of(page, size)),
             problem -> ProblemDto.map(problem)
         );
+    }
+
+    @Transactional
+    public void increaseSentSubmissions(Integer problemId) {
+        if (problemId == null) {
+            log.warn("Can't increase 'sent submissions' counter, problemId  is null.");
+            return;
+        }
+            
+        problemRepository
+            .findById(problemId)
+            .ifPresentOrElse(
+                problem -> {
+                    problem.setSentSubmissions(
+                        problem.getSentSubmissions() + 1
+                    );
+                    problemRepository.save(problem);
+                },
+                () ->
+                    log.warn(
+                        "Can't increase 'sent submissions' counter, problem with id -> {} has not been found.",
+                        problemId
+                    )
+            );
+    }
+
+    @Transactional
+    public void increaseAcceptedSubmissions(Integer problemId) {
+        if (problemId == null) {
+            log.warn("Can't increase 'accepted submissions' counter, problemId  is null.");
+            return;
+        }
+        
+        problemRepository
+            .findById(problemId)
+            .ifPresentOrElse(
+                problem -> {
+                    problem.setAcceptedSubmissions(
+                        problem.getAcceptedSubmissions() + 1
+                    );
+                    problemRepository.save(problem);
+                },
+                () ->
+                    log.warn(
+                        "Can't increase 'accepted submissions' counter, problem with id -> {} has not been found.",
+                        problemId
+                    )
+            );
     }
 }
